backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai>=1.60.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pytest>=7.4.0,<8
pytest-asyncio>=0.23.0,<0.24
ruff>=0.3.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
email-validator>=2.0
weasyprint>=60.0
jinja2>=3.1.0
